First commit with 3 APIs
Initial commit with login, register,post donation API, cors feature enabled and ADO.NET data model added

+++ CharitAble-current/Models/SP_SelectAllUsers_Result.cs
@@ -20,7 +20,7 @@ namespace CharitAble_current.Models
         public string Username { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
-        public Nullable<decimal> ContactNumber { get; set; }
+        public Nullable<long> ContactNumber { get; set; }
         public Nullable<short> VerificationCode { get; set; }
         public Nullable<bool> IsEmailVerified { get; set; }
         public byte[] ProfilePicture { get; set; }
